requirements.txt
pdfplumber==0.8.1
dateparser==1.1.8
pytest==7.4.0
openai==2.5.0
python-dotenv==1.0.0
ics==0.7.2